requirements.txt
asgiref==3.9.1
Brotli==1.1.0
cffi==1.17.1
crispy-bootstrap5==2025.6
cssselect2==0.8.0
dj-database-url==3.1.0
Django==5.2.4
django-auditlog==3.3.0
django-crispy-forms==2.4
fonttools==4.59.0
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.2.3.post1
